# Add `WP_Object_Cache` to the `gutenberg_get_global_settings` method