@oandregal 2022-10-28
@ramonjd @andrewserong @scruffian @tellthemachines @noisysocks I'm working on this PR and the only test failures are related to `WP_Block_Supports_Typography_Test`, introduced https://github.com/WordPress/gutenberg/pull/44764 and https://github.com/WordPress/gutenberg/pull/44762.

I need to look at them deeply, but thought you may be able to offer some quick thoughts.