@ramonjd 2022-10-28
> @ramonjd @andrewserong @scruffian @tellthemachines @noisysocks I'm working on this PR and the only test failures are related to WP_Block_Supports_Typography_Test, introduced https://github.com/WordPress/gutenberg/pull/44764 and https://github.com/WordPress/gutenberg/pull/44762.

I can look more closely next week, but does this PR changes the way we resolve the theme.json for the test theme we're using in the typography tests?

https://github.com/WordPress/gutenberg/blob/trunk/phpunit/data/themedir1/block-theme-child-with-fluid-typography/theme.json

The tests are failing because `"fluid": true` isn't being read (or doesn't exist)

```json
		"typography": {
			"fluid": true
		}
```