@oandregal 2022-10-31
Some initial notes to invalidate the cache for settings & styles:

- Core data changes: when WordPress is updated?
- Block data changes & block registered:
  - Different blocks may be registered across requests or across different moments in time in the same request
  - What about 3rd parties filtering the block settings?
- Theme data changes: theme switch, theme preview, theme updated
- User data changes: wp_global_styles is saved, data is imported
- Listeners for existing filters change (theme.json data or block registration)
  - Plugins installed/uninstalled/updated.
  - Plugins updated: can modify the listeners
  - Drop-in files: how to detect this if we don't keep a list of listeners?