@oandregal 2022-11-22
I've extracted two small PRs that we can fast-track and land separately:

- https://github.com/WordPress/gutenberg/pull/45971
- https://github.com/WordPress/gutenberg/pull/45969

The goal of this PR should be simply adding object cache to this method.