@spacedmonkey 2022-11-24
This is confusing to me. It seem like `gutenberg_get_global_stylesheet` and `gutenberg_get_global_settings` are caching the same data. 

The piece that needs to be cached is `WP_Theme_JSON_Resolver_Gutenberg::get_merged_data();`. That is the expensive part that needs caching. Caching twice is wasteful is likely that the cache will get out of date or not correctly be invalidated. 

Can be look at adding the caching to `WP_Theme_JSON_Resolver_Gutenberg`?