@oandregal 2022-11-24
> This is confusing to me. It seem like gutenberg_get_global_stylesheet and gutenberg_get_global_settings are caching the same data.
> 
> The piece that needs to be cached is WP_Theme_JSON_Resolver_Gutenberg::get_merged_data();. That is the expensive part that needs caching. Caching twice is wasteful is likely that the cache will get out of date or not correctly be invalidated.
> 
> Can be look at adding the caching to WP_Theme_JSON_Resolver_Gutenberg?

We talked about this at length in a recent private conversation, as well as in a few other PRs 🙂

I'm happy to discuss the direction, though at some point we need to move forward with something. If we can't agree on direction, we need to [disagree and commit](https://en.wikipedia.org/wiki/Disagree_and_commit). Being caught in an eternal loop is not fun for anyone.

We already have [caching](https://github.com/WordPress/wordpress-develop/blob/d2f420493d87465c323190892d4e7a6e4c6070a7/src/wp-includes/class-wp-theme-json-resolver.php#L642) in `WP_Theme_JSON_Resolver` and it does not work well anymore, given all the datum stored in `theme.json` have different invalidation characteristics. This was one of the [major issues I saw in the WordPress 6.1](https://github.com/WordPress/gutenberg/issues/44772) and the reason I created [this track issue](https://github.com/WordPress/gutenberg/issues/45171) to improve the situation in a strategic way (big wins first, smaller wins later).