@spacedmonkey 2022-11-25
@oandregal Can you explain why we are caching the same data twice? What is the benefit of this? The reason it keeps coming up, I don’t feel like you have explained yourself in enough detail that I feel like I can drop this. Caching the same data twice, is strange and makes cache invalidation harder, you have two caches to worry about. Not to mention, more memory used in both object cache and in persistent object caches. 

As for the fact we discussed this privately, that is part of the reason I am mentioning it again, so we can discuss, with other involved the reasoning behind this code. This conversion is public and future developers will benefit it being documented.