@spacedmonkey 2022-11-25
> Absolutely. I've been doing for a month now across many issues, so I've got some practice 🙂

I am just catching up on all those issues. There is no one place I feel like you have explained this issue in detail. So I am still trying to understand your thinking. 

To zoom out a little, the performance issue is in `WP_Theme_JSON_Resolver::get_merged_data`. Merge of the data from different sources ( even through they are cached ) results in lots of wasted compute to merge arrays together. See this screenshot from blackfire. 26% of the page load is just this one method alone. This calls theme_json, 500+ times. 

<img width="1904" alt="Screenshot 2022-11-25 at 10 31 51" src="https://user-images.githubusercontent.com/237508/203965370-b6ae3ec6-cd5f-4768-896a-6b8e9b0fc9c2.png">

Lots of work was done by @aristath to make theme json performance faster. See https://github.com/WordPress/wordpress-develop/pull/3555. 

So the core issue of performance is where we should focus our work. This is adding a cache, upstream from the core issue. Like putting a band aid over the issue, instead of treating the larger issue. 

The point of the cache, is stop the expensive compute in `WP_Theme_JSON_Resolver::get_merged_data`. Instead of caching there, we are adding two caches, which will fill up the cache. The fact that both caches are invalidated at the same time, proves that they need to be one cache. 

Let's focus on the core issue and not to cover it up with caches.